pyproject.toml
[project]
name = "sentinelx402"
version = "0.1.0"
description = "Cyber threat intelligence APIs for AI agents, paid via x402 micropayments"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.19.0",
    "greenlet>=3.0",
    "pydantic-settings>=2.0",
    "slowapi>=0.1.9",
    "nvdlib>=0.7.0",
    "httpx>=0.24.0",
]

[project.optional-dependencies]
production = [
    "gunicorn>=21.0",
    "asyncpg>=0.28.0",
    "redis>=5.0",
    "alembic>=1.12.0",
]
x402 = [
    "x402[fastapi,evm]>=0.1.0",
]
redis = [
    "redis>=5.0",
]
postgres = [
    "asyncpg>=0.28.0",
    "alembic>=1.12.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
